refactor: Separate front-end and core code.
