+++ backends/include/backend.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "registers.h"
+#include "registers.hpp"
 
 class Backend
 {
