+++ core/include/version.hpp
@@ -1,0 +1,3 @@
+#ifndef GIT_VERSION
+#define GIT_VERSION "unknown"
+#endif
