--- version.h
@@ -1,3 +1,0 @@
-#ifndef GIT_VERSION
-	#define GIT_VERSION "unknown"
-#endif
